grader-vision-update/requirements.txt
# Web Framework
fastapi==0.127.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.7.1

# Google Cloud & Gmail
google-api-python-client==2.185.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud-pubsub==2.34.0
google-cloud-storage==2.18.0

# Database (Supabase/PostgreSQL)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.31.0
psycopg2-binary==2.9.10
greenlet==3.1.1

# Document Processing
python-docx==1.2.0
PyPDF2==3.0.1
reportlab==4.4.5
Pillow==12.0.0
python-bidi==0.6.7

# NEW: PDF to Image conversion (requires poppler-utils on system)
pdf2image==1.17.0

# AI/LLM
openai==2.14.0
langchain==1.2.0
langchain-openai==1.1.6
langgraph==1.0.4

# Utilities
python-multipart==0.0.20
aiofiles==25.1.0
python-dotenv==1.2.1
httpx==0.28.1
langsmith==0.5.1

# Authentication
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-jose[cryptography]==3.3.0
